Refactor: Entity 추가에 따른 문제 관련 API 수정
MatchingService의 API들 증 Question 관련, usercode랑 Mastercode FK 참조하게 수정

Fixes: S09P12A107-135

+++ backend/src/main/java/com/ssafy/solive/api/request/QuestionDeletePutReq.java
@@ -7,5 +7,5 @@ public class QuestionDeletePutReq {
 
     Long questionId;
 
-    int studentId;
+    Long studentId;
 }
